refactor(qview): separate requests process
