+++ scripts/create_qgis_view.py
@@ -2,14 +2,13 @@
 """This script creates a PackO client view for QGIS"""
 
 import argparse
-import json
 import re
 import ast
 import os.path
 from copy import deepcopy
 from lxml import etree
-import requests
 from osgeo import gdal
+from process_requests import check_get_post, xml_from_wmts
 
 
 def read_args():
@@ -40,35 +39,6 @@ def read_args():
         print('\nArguments: ', argum)
 
     return argum
-
-
-def check_get_post(req, is_get=True):
-    """ Check GET or POST request """
-    try:
-        if is_get:
-            response = requests.get(req)
-        else:
-            response = requests.post(req)
-        response.raise_for_status()
-    except Exception as err:
-        msg = 'ERROR:\n\t'
-        try:
-            msg += f'{response.json()}\n\t'
-        except json.decoder.JSONDecodeError:
-            pass
-        msg += str(err)
-        raise SystemExit(msg)
-    return response
-
-
-def xml_from_wmts(wmts_in, xml_out):
-    """ Create xml file from wmts GetCapabilities request """
-    src_ds = gdal.Open(wmts_in)
-    _ = gdal.Translate(xml_out, src_ds, format='WMTS')
-    # close dataset to flush to disk
-    _ = None
-    src_ds = None
-    print(f"File '{xml_out}' written")
 
 
 def suppress_cachetag(xml_in, xml_out):
